wip: add move with xr priority
